recognition/requirements.txt
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
pillow==11.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==70.2.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
typing_extensions==4.12.2
